requirements.txt
langchain==0.3.3
langchain-ollama==0.2.0
langgraph==0.2.39
matplotlib==3.9.2
ollama==0.3.3
pandas==2.2.3
requests==2.32.3
seaborn==0.13.2
fpdf2==2.8.1
pytest==8.3.3
xgboost==2.1.2
scikit-learn==1.5.2
lightgbm==4.5.0
imbalanced-learn==0.12.4
sdv==1.17.1
numpy==1.26.4
shap==0.46.0